+++ ChangeLog.md
@@ -6,6 +6,8 @@
 - Add vimrc to load the plugin "vundle" and set base configuration.
 - Add plugin "NERDTree" to show file expolor in vim.
 - Add plugin "tagbar" to show the tags in vim for C/C++ code files.
+- Add the plugin "YouCompleteMe" to help to edit code files.
+- Add shell file to use this project to setup vim.
 
 ### Changed
 - Update the template for git-commit command.